+++ app/src/main/java/com/org/marton/studio/project/eldarwallet/ui/activities/qrpay/adapter/DigitalCardQrAdapter.kt
@@ -39,7 +39,7 @@ class DigitalCardQrAdapter(
             holder.itemView.setBackgroundColor(ContextCompat.getColor(holder.itemView.context, R.color.default_card_color)) // Restaura el color de fondo
         }
         holder.cardBankNameTextView.text = CardUtils.getBankNameByCode(digitalCard.bank.toInt())
-        holder.cardNumberTextView.text = CardUtils.formatCardNumber(digitalCard.number)
+        holder.cardNumberTextView.text = CardUtils.formatCardNumber(digitalCard.number.toLong())
         holder.cardManagmentNameTextView.text = CardUtils.getBrandCardNameByCode(digitalCard.brand)
         holder.cardTypeTextView.text = CardUtils.getTypeCardDescByCode(digitalCard.type)
         holder.itemView.setOnClickListener {
